requirements.txt
fastapi
uvicorn
firebase-admin
qrcode
python-multipart
jinja2
pydantic
python-dotenv
Pillow
